requirements.txt
flask==3.1.3
gunicorn==25.1.0
supabase==2.28.2
python-dotenv==1.2.2
requests==2.32.5
pyjwt[crypto]==2.12.1
